feat: #200 🎸 headerをコンポーネント化
